Move nnue evaluation stuff from evaluate.h to nnue/evaluate_nnue.h
