+++ src/learn/gensfen.cpp
@@ -12,6 +12,7 @@
 
 #include "extra/nnue_data_binpack_format.h"
 
+#include "nnue/evaluate_nnue.h"
 #include "nnue/evaluate_nnue_learner.h"
 
 #include "syzygy/tbprobe.h"
